Re-arrange all img, moved them to img folder.
