+++ .gitignore
@@ -22,17 +22,14 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
+#firebase
+.firebase
+
 #api keys
 .env
-firestore-algolia-search.env
 sa.json
-//newstimeline-62758-firebase-adminsdk-de5mn-531e006ab4.json//舊的
 newstimeline-62758-firebase-adminsdk-de5mn-6476ca6af7.json
 //app engine在用的
 newstimeline-62758-461ddc6816ff.json
-algoliaApiKey.tsx
 
-#content crawling .js
-api.js
-fetchContent.js
-firebase_fetchNews.js
+
