+++ src/img/savedSign.png
